Decoupled CordaVersionProvider. Moved common files to common-logging to lower dependencies on the node explorer.
