+++ common/logging/build.gradle
@@ -1,0 +1,26 @@
+apply plugin: 'kotlin'
+
+apply plugin: 'net.corda.plugins.publish-utils'
+apply plugin: 'com.jfrog.artifactory'
+
+dependencies {
+    compile group: "org.jetbrains.kotlin", name: "kotlin-stdlib-jdk8", version: kotlin_version
+    compile group: "org.jetbrains.kotlin", name: "kotlin-reflect", version: kotlin_version
+
+    compile group: "com.typesafe", name: "config", version: typesafe_config_version
+    compile "org.apache.logging.log4j:log4j-slf4j-impl:$log4j_version"
+
+
+    compile "com.jcabi:jcabi-manifests:$jcabi_manifests_version"
+
+
+    testCompile project(":test-utils")
+}
+
+jar {
+    baseName 'corda-common-logging'
+}
+
+publish {
+    name jar.baseName
+}
