+++ node/build.gradle
@@ -73,6 +73,7 @@ dependencies {
     compile project(':tools:cliutils')
     compile project(':common-validation')
     compile project(':common-configuration-parsing')
+    compile project(':common-logging')
     
     // Backwards compatibility goo: Apps expect confidential-identities to be loaded by default.
     // We could eventually gate this on a target-version check.
