+++ settings.gradle
@@ -70,6 +70,10 @@ project(":common-validation").projectDir = new File("$settingsDir/common/validat
 
 include 'common-configuration-parsing'
 project(":common-configuration-parsing").projectDir = new File("$settingsDir/common/configuration-parsing")
+
+
+include 'common-logging'
+project(":common-logging").projectDir = new File("$settingsDir/common/logging")
 // Common libraries - end
 
 apply from: 'buildCacheSettings.gradle'
